Workaround for USD Issue 2013
See https://github.com/PixarAnimationStudios/USD/issues/2013

I do hope to get a better workaround either in the issue comments or on
this USD-interests thread:

https://groups.google.com/g/usd-interest/c/5LZT-IOaOTU/m/XwLzFTq7AgAJ

But, in the meantime, creating a dummy prim under the material is enough to
cause the resync to happen... Feel free to revoke my coding license.
This is ugly.