Update exporters for string varnames
The varname input of all UsdPrimVarReader nodes went from a Token to a
String in USD 20.11. This updates the exporter to follow the convention
currently in use, and allows re-using varnames with MaterialX.

We also do not hardcode the output name of MaterialX nodes. We query the
Sdr registry for the correct information.

There are still too many hardcoded names and types in the exporter. To be fixed later.